site/requirements.txt
MSApi>=0.2.11
Django==6.0.3
pandas>=1.1.5
ruff>=0.8.3
pydantic>=2.10.4
requests>=2.32.3
gunicorn>=20.0.4
django-colorful>=1.3
ratelimit>=2.2.1
django_extra_views>=0.16.0
ortools>=9.15.6755